+++ react-components/src/components/Architecture/TreeView/utilities/constants.tsx
@@ -1,0 +1,20 @@
+/*!
+ * Copyright 2024 Cognite AS
+ */
+
+/* eslint-disable react/prop-types */
+
+export const TEXT_COLOR = 'black';
+export const DISABLED_TEXT_COLOR = 'lightgray';
+export const SELECTED_TEXT_COLOR = 'white';
+export const SELECTED_BACKGROUND_COLOR = 'highlight';
+export const HOVER_TEXT_COLOR = 'black';
+export const HOVER_BACKGROUND_COLOR = 'lightgray';
+export const CARET_COLOR = 'gray';
+export const HOVER_CARET_COLOR = 'highlight';
+export const CARET_SIZE = 20;
+export const GAP_TO_CHILDREN = 16;
+export const GAP_BETWEEN_ITEMS = 4;
+export const LOADING_LABEL = 'Loading ...';
+export const LOAD_MORE_LABEL = 'Load more ...';
+export const BACKGROUND_COLOR = 'white';
